Casts size type to int

+++ src/lex.cpp
@@ -8,7 +8,7 @@ int main() {
     Lexer lexer = Lexer("code.txt");
     
     std::vector<Token> seq= lexer.lexer();
-    for(int i = 0; i < seq.size(); i++){
+    for(int i = 0; i < (int)(seq.size()); i++){
         std::cout << seq.at(i).line << " " << seq.at(i).column << " " << seq.at(i).token << std::endl;
     }
     return 0;
